+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Connecteur i-Parapheur : Ajout d'une action pour vider le cache WSDL #464 
 - Page d'informations suplémentaires sur les travaux regroupés par verrou et par états #459
 - Françisation et homogénéisation des noms des connecteur bouchon #466
+- Il est maintenant nécessaire de saisir les informations complémentaires pour l'envoi direct en GED sans passage par le TDT (actes, helios) #481 #437   
+- Les TDT peuvent maintenant utiliser Pastell pour le versement GED   
 
 ## Ajouts